+++ Source/ImoutoRebirth.Room/ImoutoRebirth.Room.Domain/ImoutoPicsUploaderState.cs
@@ -1,0 +1,10 @@
+﻿namespace ImoutoRebirth.Room.Domain;
+
+public class ImoutoPicsUploaderState
+{
+    public bool IsEnabled { get; private set; } = true;
+
+    public void Enable() => IsEnabled = true;
+    
+    public void Disable() => IsEnabled = false;
+}