+++ Source/ImoutoRebirth.Room/ImoutoRebirth.Room.UI/WebApi/WebApiStartup.cs
@@ -46,6 +46,7 @@ public static class WebEndpointsExtensions
         app.MapCollectionFilesEndpoints();
         app.MapDestinationFoldersEndpoints();
         app.MapSourceFoldersEndpoints();
+        app.MapImoutoPicsUploaderEnabled();
         
         return app;
     }